requirements.txt
asgiref==3.8.1
Django==5.2.1
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.9
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.13.2
